### 2020-10-09 #2736 @whedon
[PRE REVIEW]: hectorui: A web-based interactive scenario builder and visualization application for the Hector climate model
**Submitting author:** @evanoffPNNL (<a href="http://orcid.org/0000-0002-8858-2783">Jason Evanoff</a>)
**Repository:** <a href="https://github.com/JGCRI/hectorui/" target ="_blank">https://github.com/JGCRI/hectorui/</a>
**Version:** 1.1.0
