### 2020-10-09 #2736 @kthyng
@evanoffPNNL Thank you for your submission! I will be your editor. Please add a "Statement of Need" to your paper to start. Also, can you recommend reviewers for this submission? You can recommend from your network or from [this list](https://bit.ly/joss-reviewers) of volunteers (as long as not conflicting). If you are providing github handles, please do not use "@" so that they are not prematurely notified with excessive email.
