+++ paper/paper.bib
@@ -1,93 +1,0 @@
-@article{Hartin2015,
-  author = {Hartin, C. A. and Patel, P. and Schwarber, A. and Link, R. P. and Bond-Lamberty, B. P.},
-  title = {A simple object-oriented and open-source model for scientific and policy analyses of the global climate system -- Hector v1.0},
-  journal = {Geoscientific Model Development},
-  volume = {8},
-  year = {2015},
-  number = {4},
-  pages = {939--955},
-  url = {http://www.geosci-model-dev.net/8/939/2015/},
-  doi = {10.5194/gmd-8-939-2015}
-}
-@article{Hartin2016,
-  author = {Hartin, C. A. and Bond-Lamberty, B. and Patel, P. and Mundra, A.},
-  title = {Ocean acidification over the next three centuries using a simple global
-  climate carbon-cycle model: projections and sensitivities},
-  journal = {Biogeosciences},
-  volume = {13},
-  year = {2016},
-  number = {15},
-  pages = {4329--4342},
-  url = {http://www.biogeosciences.net/13/4329/2016/},
-  doi = {10.5194/bg-13-4329-2016}
-}
-@article{snyder2019joint,
-  title={Joint emulation of Earth System Model temperature-precipitation realizations with internal variability and space-time and cross-variable correlation: fldgen v2. 0 software description},
-  author={Snyder, Abigail and Link, Robert and Dorheim, Kalyn and Kravitz, Ben and Bond-Lamberty, Ben and Hartin, Corinne},
-  journal={Plos one},
-  volume={14},
-  number={10},
-  pages={e0223542},
-  year={2019},
-  publisher={Public Library of Science San Francisco, CA USA}
-}
-@article{calvin2019gcam,
-  title={GCAM v5. 1: representing the linkages between energy, water, land, climate, and economic systems},
-  author={Calvin, Katherine and Patel, Pralit and Clarke, Leon and Asrar, Ghassem and Bond-Lamberty, Ben and Cui, Ryna Yiyun and Di Vittorio, Alan and Dorheim, Kalyn and Edmonds, Jae and Hartin, Corinne and others},
-  journal={Geoscientific Model Development (Online)},
-  volume={12},
-  number={PNNL-SA-137098},
-  year={2019},
-  publisher={Lawrence Berkeley National Lab.(LBNL), Berkeley, CA (United States); Pacific~…}
-}
-@article{riahi2007scenarios,
-  title={Scenarios of long-term socio-economic and environmental development under climate stabilization},
-  author={Riahi, Keywan and Gr{\"u}bler, Arnulf and Nakicenovic, Nebojsa},
-  journal={Technological Forecasting and Social Change},
-  volume={74},
-  number={7},
-  pages={887--935},
-  year={2007},
-  publisher={Elsevier}
-}
-@article{riahi2011rcp,
-  title={RCP-8.5: exploring the consequence of high emission trajectories},
-  author={Riahi, Keywan and Krey, V and Rao, S and Chirkov, V and Fischer, G and Kolp, P and Kindermann, G and Nakicenovic, N and Rafai, P},
-  journal={Climatic Change. doi},
-  volume={10},
-  pages={1007},
-  year={2011}
-}
-@article{fujino2006multi,
-  title={Multi-gas mitigation analysis on stabilization scenarios using AIM global model},
-  author={Fujino, Junichi and Nair, Rajesh and Kainuma, Mikiko and Masui, Toshihiko and Matsuoka, Yuzuru},
-  journal={The Energy Journal},
-  number={Special Issue\# 3},
-  year={2006},
-  publisher={International Association for Energy Economics}
-}
[…]
