+++ tests/testthat.R
@@ -1,4 +1,0 @@
-library(testthat)
-library(hectorui)
-
-test_check("hectorui")
