+++ R/hectorui.R
@@ -1,2 +1,0 @@
-
-"_PACKAGE"

[…]
