### 2020-10-09 #2736 @crvernon
@kthyng @evanoffPNNL The latest release is:  https://github.com/JGCRI/hectorui/releases/tag/v1.1.1 and the accompanying DOI from Zenodo link is:   http://doi.org/10.5281/zenodo.4075376

Thanks!
