### 2020-10-10 #2736 @kthyng
Hi @crvernon! We'll need that information at the end of the review process to make sure everything is up to date at that point, and there should be an up to date zenodo archive at that time too.
