### 2020-10-19 f0ed022
Updated bib file to fix missing DOI
