+++ paper/paper.bib
@@ -29,7 +29,8 @@
   number={10},
   pages={e0223542},
   year={2019},
-  publisher={Public Library of Science San Francisco, CA USA}
+  publisher={Public Library of Science San Francisco, CA USA},
+  doi = {https://doi.org/10.1371/journal.pone.0223542}
 }
 @article{calvin2019gcam,
   title={GCAM v5. 1: representing the linkages between energy, water, land, climate, and economic systems},
@@ -38,7 +39,8 @@
   volume={12},
   number={PNNL-SA-137098},
   year={2019},
-  publisher={Lawrence Berkeley National Lab.(LBNL), Berkeley, CA (United States); Pacific~…}
+  publisher={Lawrence Berkeley National Lab.(LBNL), Berkeley, CA (United States); Pacific~…},
+  doi = {https://doi.org/10.5194/gmd-12-677-2019}
 }
 @article{riahi2007scenarios,
   title={Scenarios of long-term socio-economic and environmental development under climate stabilization},
@@ -48,7 +50,8 @@
   number={7},
   pages={887--935},
   year={2007},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {https://doi.org/10.1016/j.techfore.2006.05.026}
 }
 @article{riahi2011rcp,
   title={RCP-8.5: exploring the consequence of high emission trajectories},
@@ -79,7 +82,8 @@
   number={1-2},
   pages={77},
   year={2011},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {https://doi.org/10.1007/s10584-011-0151-4}
 }
 @article{van2011rcp2,
   title={RCP2. 6: exploring the possibility to keep global mean temperature increase below 2 C},
@@ -89,5 +93,6 @@
   number={1-2},
   pages={95},
   year={2011},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {10.1007/s10584-011-0152-3}
 }
