+++ paper/paper.bib
@@ -59,7 +59,8 @@
   journal={Climatic Change. doi},
   volume={10},
   pages={1007},
-  year={2011}
+  year={2011},
+  doi = {10.1007/s10584-011-0149-y}
 }
 @article{fujino2006multi,
   title={Multi-gas mitigation analysis on stabilization scenarios using AIM global model},
@@ -67,7 +68,8 @@
   journal={The Energy Journal},
   number={Special Issue\# 3},
   year={2006},
-  publisher={International Association for Energy Economics}
+  publisher={International Association for Energy Economics},
+  doi = {10.5547/ISSN0195-6574-EJ-VolSI2006-NoSI3-17}
 }
 @article{clarke2007scenarios,
   title={Scenarios of greenhouse gas emissions and atmospheric concentrations},
