### 2020-10-19 cf1d845
updating DOI to remove http prefix
