### 2020-10-19 b594c6b
Removed doi prefix in references
