+++ paper/paper.bib
@@ -30,7 +30,7 @@
   pages={e0223542},
   year={2019},
   publisher={Public Library of Science San Francisco, CA USA},
-  doi = {doi.org/10.1371/journal.pone.0223542}
+  doi = {10.1371/journal.pone.0223542}
 }
 @article{calvin2019gcam,
   title={GCAM v5. 1: representing the linkages between energy, water, land, climate, and economic systems},
@@ -40,7 +40,7 @@
   number={PNNL-SA-137098},
   year={2019},
   publisher={Lawrence Berkeley National Lab.(LBNL), Berkeley, CA (United States); Pacific~…},
-  doi = {doi.org/10.5194/gmd-12-677-2019}
+  doi = {10.5194/gmd-12-677-2019}
 }
 @article{riahi2007scenarios,
   title={Scenarios of long-term socio-economic and environmental development under climate stabilization},
@@ -51,7 +51,7 @@
   pages={887--935},
   year={2007},
   publisher={Elsevier},
-  doi = {doi.org/10.1016/j.techfore.2006.05.026}
+  doi = {10.1016/j.techfore.2006.05.026}
 }
 @article{riahi2011rcp,
   title={RCP-8.5: exploring the consequence of high emission trajectories},
@@ -85,7 +85,7 @@
   pages={77},
   year={2011},
   publisher={Springer},
-  doi = {doi.org/10.1007/s10584-011-0151-4}
+  doi = {10.1007/s10584-011-0151-4}
 }
 @article{van2011rcp2,
   title={RCP2. 6: exploring the possibility to keep global mean temperature increase below 2 C},
