### 2020-10-26 #2736 @kthyng
@marvinjonathcn: are you interested in reviewing this submission to JOSS? Reviews for JOSS are open on github and are checklist-driven. As you find items to bring up in the review process, you open issues to the software repository itself, link to the review issue, and explain the comment. We're asking for reviews within 5-6 weeks, a long timeline due to the pandemic. You can find more information about the reviewing process here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#
