### 2020-10-26 #2736 @nuest
@kthyng Sorry, I am just about to start another JOSS review and wouldn't be able to get to this even within the extended timeline.

@tscheypidi are you a JOSS reviewer yet?
