### 2020-10-26 #2736 @kthyng
@tscheypidi [Here](https://joss.theoj.org/reviewer-signup.html) is the link to sign up as a JOSS reviewer generally if you'd like. Are you interested in being a reviewer on this submission?
