### 2020-10-27 #2782 @kthyng
@marvinjonathcn, @tscheypidi: this is where the actual review takes places. Please look through [reviewer docs](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) to see how the review process works, and look through the checklist above too. As issues arise in the review process, open an issue per comment in the submission repo on github and link back to this issue. Don't hesitate to ask if you have any questions about how it all works especially since it is a little unusual!
