### 2020-10-29 #2782 @tscheypidi
@evanoffPNNL, I am just checking the list of authors and did not find all mentioned authors in the commit history of hectorui. Could you please clarify how Stephanie Waldhoff, Abigail Snyder and Corinne Hartin contributed to the Software?
