### 2020-10-29 JGCRI/hectorui#19 @tscheypidi
Missing installation instructions
Hi @evanoffPNNL,

I am currently looking for installation instructions how to install and run hectorui on my local system. Could you please add some instructions either to the README.md or tutorial file (how to install it & how to run the shiny app locally)?

This is related to https://github.com/openjournals/joss-reviews/issues/2782.

Thanks, Jan
