### 2020-10-29 JGCRI/hectorui#20 @tscheypidi
Missing unit information
The custom emissions section does not seem to tell in which unit the custom emissions value is expected (Tg?). Please add this information.

(related to https://github.com/openjournals/joss-reviews/issues/2782)
