### 2020-10-29 JGCRI/hectorui#21 @tscheypidi
Somewhat unclear error message for unreasonable parameter settings
When setting a parameter to an unreasonable value It always seems to trigger the same error message:

```
Error: Error in hector::run(hcores[[i]], globalVars[["endDate"]]): Error while running hector: msg: Assertion failed: bad Tk value
func: ocean_csys_run
file: ocean_csys.cpp
ffile: ocean_csys.cpp

line: 130
```
Examples which trigger this error are (while keeping all other parameters to default settings):

1. Setting arosol scaling factor to 100
2. setting custom emissions for 2000 -> 2010 to 1000

Is it possible to return in these instances a more meaningful error message to better communicate to less experienced users that they probably made an inconsistent choice of parameters?

(This is related to https://github.com/openjournals/joss-reviews/issues/2782 but only meant as suggestion)
