### 2020-10-29 JGCRI/hectorui#22 @tscheypidi
Cite software used in hectorui
Please cite the shiny package in your paper (`citation("shiny")` can be used in R to retrieve the citation). If there are further software or software packages hectorui heavily relies on, but which is not mentioned/cited at the moment, please cite it as well.

Linked to: https://github.com/openjournals/joss-reviews/issues/2782
