### 2020-10-29 JGCRI/hectorui#23 @tscheypidi
Expand Contact/Contribution information in README
Please expand a bit the information on how one could contribute to hectorui or get in contact with you in case of problems. In particular I am looking for answers to the following questions:

- Who exactly should I contact, if I want to collaborate / How should I contact you?
- If I need support, should I contact you directly/via mail or should I open an issue in the repository or is there a community I could post my questions to?
- Should issues/bugs reported via the issue tracker or differently?

https://github.com/openjournals/joss-reviews/issues/2782
