### 2020-11-03 #2782 @tscheypidi
I am currently waiting for installation instructions (https://github.com/JGCRI/hectorui/issues/19) which are required in order to check installation and functionality of the software.

All other parts have been checked and will get ticked off as soon as the corresponding issues have been resolved (https://github.com/JGCRI/hectorui/issues/20, https://github.com/JGCRI/hectorui/issues/22 and https://github.com/JGCRI/hectorui/issues/23).

https://github.com/JGCRI/hectorui/issues/21 is somehow independent as it is a suggestion but no requirement to complete the review process.
