### 2020-11-09 fb53d67
Fixes requested issues by reviewer -  installation instructions, unit information and limits, readme
