+++ paper/paper.bib
@@ -91,3 +91,10 @@
   year={2011},
   publisher={Springer}
 }
+@Manual{,
+    title = {shiny: Web Application Framework for R},
+    author = {Winston Chang and Joe Cheng and JJ Allaire and Yihui Xie and Jonathan McPherson},
+    year = {2020},
+    note = {R package version 1.4.0.2},
+    url = {https://CRAN.R-project.org/package=shiny},
+  }
