### 2020-11-09 7a92878
Revert "Fixes requested issues by reviewer -  installation instructions, unit information and limits, readme"
This reverts commit fb53d67541794e79cfda2f4b8eca5eda4fc77c1f.
