+++ README.md
@@ -6,15 +6,14 @@ A web-based interactive scenario builder and visualization application for the H
 
 ## Using `hectorui`
 
-Just getting started with `hectorui`?  We have constructed a tutorial to examine a sample use-cases available here:  [`hectorui` Tutorial](https://rpubs.com/jevanoff/hectorui)
+Just getting started with `hectorui`?  We have constructed a tutorial to examine some sample use-cases which are available here:  [Tutorial](https://jgcri.github.io/hectorui/articles/Tutorial.html)
 
-To navigate directly to the `hectorui` app:  [`hector-ui` Application](https://jgcri.shinyapps.io/HectorUI/)
+To navigate directly to the `hectorui` app **CLICK the following image**:
 
-![`hectorui` map scenario interface](https://raw.githubusercontent.com/JGCRI/hectorui/master/paper/figure1.png)
+[![`hectorui` map scenario interface](https://raw.githubusercontent.com/JGCRI/hectorui/master/paper/figure1.png)](https://jgcri.shinyapps.io/HectorUI/)
 
-_The map scenario interface to `hectorui` that allows the user to parameterize inputs and visualize run outputs interactively._
-
+_The map scenario interface to `hectorui` that allows the user to parameterize inputs and visualize run outputs interactively.  Image contains a link to the application._
 
 ## Contributing to `hectorui`
 
-We welcome contributions to `hectorui` from the development community.  Please contact us if you want to collaborate!  The `hectorui` GitHub repository is accessible here:  [`hector-ui` GitHub Repository](https://github.com/JGCRI/hectorui)
+We welcome contributions to `hectorui` from the development community.  Please contact us if you want to collaborate!  The `hectorui` GitHub repository is accessible here:  [GitHub Repository](https://github.com/JGCRI/hectorui)
