+++ paper/paper.bib
@@ -98,3 +98,10 @@
   publisher={Springer},
   doi = {10.1007/s10584-011-0152-3}
 }
+@Manual{,
+  title = {shiny: Web Application Framework for R},
+  author = {Winston Chang and Joe Cheng and JJ Allaire and Yihui Xie and Jonathan McPherson},
+  year = {2020},
+  note = {R package version 1.4.0.2},
+  url = {https://CRAN.R-project.org/package=shiny},
+}
