+++ inst/shinyApp/parameters.r
@@ -1,6 +1,6 @@
 # Contains parameter related functions and observers
 
-#' Assign Parameters from numeric components to core
+#' Assign Parameters from model to core
 #'
 #' @return no return value
 #' @export
@@ -8,13 +8,13 @@
 #' @examples
 setCoreParameters <- function(hcore)
 {
-    hector::setvar(hcore, dates = NA, var = globalParameters['aero'], values = c(as.double(paramsList[['alpha']])), unit = "unitless")
-    hector::setvar(hcore, dates = NA, var = globalParameters['beta'], values = c(as.double(paramsList[['beta']])), unit = "unitless")
-    hector::setvar(hcore, dates = NA, var = globalParameters['diff'], values = c(as.double(paramsList[['diff']])), unit = "cm2/s")
-    hector::setvar(hcore, dates = NA, var = globalParameters['ecs'],  values = c(as.double(paramsList[['S']])), unit = "degC")
-    hector::setvar(hcore, dates = NA, var = globalParameters['pco2'], values = c(as.double(paramsList[['C']])), unit = "ppmv CO2")
-    hector::setvar(hcore, dates = NA, var = globalParameters['q10'],  values = c(as.double(paramsList[['q10_rh']])), unit = "unitless")
-    hector::setvar(hcore, dates = NA, var = globalParameters['volc'], values = c(as.double(paramsList[['volscl']])), unit = "unitless")
+  hector::setvar(hcore, dates = NA, var = globalParameters['aero'], values = c(as.double(paramsList[['alpha']])), unit = "unitless")
+  hector::setvar(hcore, dates = NA, var = globalParameters['beta'], values = c(as.double(paramsList[['beta']])), unit = "unitless")
+  hector::setvar(hcore, dates = NA, var = globalParameters['diff'], values = c(as.double(paramsList[['diff']])), unit = "cm2/s")
+  hector::setvar(hcore, dates = NA, var = globalParameters['ecs'],  values = c(as.double(paramsList[['S']])), unit = "degC")
+  hector::setvar(hcore, dates = NA, var = globalParameters['pco2'], values = c(as.double(paramsList[['C']])), unit = "ppmv CO2")
+  hector::setvar(hcore, dates = NA, var = globalParameters['q10'],  values = c(as.double(paramsList[['q10_rh']])), unit = "unitless")
+  hector::setvar(hcore, dates = NA, var = globalParameters['volc'], values = c(as.double(paramsList[['volscl']])), unit = "unitless")
 }
 
 #' Assign Parameters from parameter model values into numeric input components
@@ -121,7 +121,7 @@ loadParameters <- function()
 {
   print("in load params")
 
-  # Fetch hector parameters from core
+  # Fetch hector parameters from core sadf asdfasdf ds
   for(i in 1:length(hcores))
   {
     hdata <- hector::fetchvars(core = hcores[[i]], dates = NA, vars = globalParameters, "\n")
@@ -161,51 +161,90 @@ setParameters <- function()
   if(length(hcores) > 0)
   {
     newVals <- vector()
+    pass_check <- TRUE
     # Run through variables and make sure none are left empty and update the top level scope paramsList variable
-    # and the hector core with any changed values
+    # and the hector core with any changed values.
     tryCatch(
       {
         for(i in 1:length(hcores))
         {
-          if(!is.na(input$input_aero))
+          if(!is.na(input$input_aero) && input$input_aero <= 1 && input$input_aero > 0)
           {
             hector::setvar(hcores[[i]], dates = NA, var = globalParameters['aero'], values = c(as.double(input$input_aero)), unit = "unitless")
             paramsList['alpha'] <<- as.double(input$input_aero)
           }
-          if(!is.na(input$input_beta))
+          else
+          {
[…]
