### 2020-11-09 JGCRI/hectorui#23 @evanoffPNNL
Added email and contact information in 1.1.4
