### 2020-11-09 JGCRI/hectorui#22 @evanoffPNNL
Added citation in 1.1.4
