### 2020-11-09 JGCRI/hectorui#21 @evanoffPNNL
Changed error message to be more friendly with suggested limits in 1.1.4
