### 2020-11-09 JGCRI/hectorui#20 @evanoffPNNL
Added (Tg) to page for clarification
