### 2020-11-09 JGCRI/hectorui#19 @evanoffPNNL
Included basic installation instructions in 1.1.4
