### 2020-11-09 JGCRI/hectorui#19 @tscheypidi
@evanoffPNNL Hi Jason, thanks for your update. What I am still missing in the README is an short explanation how to run the app locally. I could start it using `shiny::runApp(system.file("shinyApp",package = "hectorui"))`, but I am not sure whether this is the envisaged method.
