### 2020-11-09 JGCRI/hectorui#22 @tscheypidi
@evanoffPNNL Please also add the citation to the paper. Currently, the citation just seems to be added to the .bib file, but not referred to in the paper itself.
