### 2020-11-10 21a38e6
Update paper to include shiny citation and info text
