+++ paper/paper.bib
@@ -98,7 +98,7 @@
   publisher={Springer},
   doi = {10.1007/s10584-011-0152-3}
 }
-@Manual{,
+@Manual{shiny,
   title = {shiny: Web Application Framework for R},
   author = {Winston Chang and Joe Cheng and JJ Allaire and Yihui Xie and Jonathan McPherson},
   year = {2020},
