+++ README.md
@@ -19,6 +19,8 @@ To install hectorui as an R package for local use, please follow these steps in 
 - library(devtools)
 - devtools::install_github("JGCRI/hectorui")
 
+To run the app after installation, the method depends on your R environment. If you are using RStudio, simply open the server.r or ui.r file and hit "> Run App". If you are working in a command line environment, start R.exe and enter the following command: shiny::runApp(system.file("shinyApp",package = "hectorui")).
+
 ## Contributing to `hectorui`
 
 We welcome contributions to `hectorui` from the development community.  Please contact us if you want to collaborate!  The `hectorui` GitHub repository is accessible here:  [`hector-ui` GitHub Repository](https://github.com/JGCRI/hectorui)
