+++ DESCRIPTION
@@ -13,6 +13,7 @@ License: GPL-3
 Encoding: UTF-8
 LazyData: true
 Imports: 
+    maps,
     remotes,
     shiny,
     shinyjs,
