### 2020-12-01 JGCRI/hectorui#26 @marvinjonathcn
 Original "HECTOR" model operating information
Hi @evanoffPNNL.

In the tutorial and in the "hectorui" paper the "Hector" model is mentioned, however in the tutorial and repository there is no additional binding link such as "See more", to obtain information about the science behind the original Hector model. How can I find more information about "Hector"?

Note: The shiny application does have information and binding links to the Hector model.

Related to : [openjournals/joss-reviews#2782](https://github.com/openjournals/joss-reviews/issues/2782).
