### 2020-12-01 JGCRI/hectorui#27 @marvinjonathcn
 R package usage examples
Hi @evanoffPNNL .

In the "hectorui" repository and tutorial, the possibility of working with the R package is mentioned. The installation is specifically detailed in the repository, however there are no codes or manuals with at least one example of using the package a Simple view and a binding "See more" link in the tutorial would be missing since its use is mentioned but there is no additional information.

Related to : [openjournals/joss-reviews#2782](https://github.com/openjournals/joss-reviews/issues/2782).
