### 2020-12-01 JGCRI/hectorui#28 @marvinjonathcn
 Download spatial information in raster format
Hi @evanoffPNNL.

Is it possible to download the information generated in the global maps in raster format? . Currently the download only allows to obtain this information in .png format

Related to : [openjournals/joss-reviews#2782](https://github.com/openjournals/joss-reviews/issues/2782).
