### 2020-12-01 JGCRI/hectorui#29 @marvinjonathcn
 Application transition from CMIP5 to CMIP6
Hi @evanoffPNNL.

Currently, the generation of models of the CMIP5 phase is outdated due to the publication of the models of the CMIP6 phase since 2019. Currently most of the data is already in the repository of [ESGF](https://esgf-node.llnl.gov/search/cmip6/) and [Google](https://console.cloud.google.com/marketplace/product/noaa-public/cmip6?filter=solution-type:dataset&q=cmip6&id=88f005b5-43c4-4698-89c9-c1a4572fbcf1).

It would be good to take it with consideration.

Related to : [openjournals/joss-reviews#2782](https://github.com/openjournals/joss-reviews/issues/2782).
