### 2020-12-01 JGCRI/hectorui#28 @evanoffPNNL
At this time we are only supporting png, but your suggestion makes a lot of sense and we will add it to the list of suggested upgrades if that is sufficient?
