### 2020-12-02 5f8a838
add local dev instructions to readme
