### 2020-12-02 aa7775a
add hector info to readme
