+++ README.md
@@ -40,3 +40,10 @@ Users familiar with R Shiny can add new features by working directly in the `hec
 We welcome contributions to `hectorui` from the development community.  Please contact us if you want to collaborate!  The `hectorui` GitHub repository is accessible here:  [`hector-ui` GitHub Repository](https://github.com/JGCRI/hectorui)
 
 For more information about contributing, please contact Jason Evanoff at jason.evanoff@pnnl.gov or Chris Vernon at chris.vernon@pnnl.gov
+
+## Learn More About Hector
+Read more about the Hector simple climate model here:  [`Hector` Documentation](https://jgcri.github.io/hector/). `Hector` has been published in the following:
+
+```
+Hartin CA, Patel P, Schwarber A, Link R, Bond-Lamberty B (2015). “A simple object-oriented and open-source model for scientific and policy analyses of the global climate system–Hector v1.0.” Geoscientific Model Development, 8(4), 939–955.
+```
