+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 1.6.1
 pkgdown_sha: ~
 articles:
   Tutorial: Tutorial.html
-last_built: 2020-10-08T14:07Z
+last_built: 2020-12-02T16:29Z
 
