### 2020-12-02 #2782 @crvernon
@kthyng ... @evanoffPNNL and I have addressed all reviewer comments and all changes are currently represented in the new release v1.1.5 (GitHub: https://github.com/JGCRI/hectorui/releases/tag/v1.1.5 ; Zenodo:  https://zenodo.org/record/4302333#.X8fCSKpKhUM).  Please let us know if further information is needed.

We have found this review to be very helpful and it has strengthened this package!  Thanks to @tscheypidi and @marvinjonathcn for all of their comments and suggestions!
