### 2020-12-04 #2782 @kthyng
@crvernon, @evanoffPNNL Can you update the metadata for your Zenodo archive so that the title and author list exactly match your JOSS paper?
