### 2020-12-04 #2782 @kthyng
@crvernon, @evanoffPNNL Your paper looks good, though please go through your references in detail and fix anything that looks off. For example, one reference has "r" (uncapitalized), and the final reference maybe should have different formatting for 2 deg C? There may be more such items. To preserve capitalization from the bib file, you can put {} around a single letter, word, or more.
