### 2020-12-08 fc65fc0
fix reference formatting in bibtex for joss paper
