+++ paper/paper.bib
@@ -1,6 +1,6 @@
 @article{Hartin2015,
   author = {Hartin, C. A. and Patel, P. and Schwarber, A. and Link, R. P. and Bond-Lamberty, B. P.},
-  title = {A simple object-oriented and open-source model for scientific and policy analyses of the global climate system -- Hector v1.0},
+  title = {A simple object-oriented and open-source model for scientific and policy analyses of the global climate system -- {H}ector v1.0},
   journal = {Geoscientific Model Development},
   volume = {8},
   year = {2015},
@@ -22,7 +22,7 @@
   doi = {10.5194/bg-13-4329-2016}
 }
 @article{snyder2019joint,
-  title={Joint emulation of Earth System Model temperature-precipitation realizations with internal variability and space-time and cross-variable correlation: fldgen v2. 0 software description},
+  title={Joint emulation of {E}arth system model temperature-precipitation realizations with internal variability and space-time and cross-variable correlation: fldgen v2.0 software description},
   author={Snyder, Abigail and Link, Robert and Dorheim, Kalyn and Kravitz, Ben and Bond-Lamberty, Ben and Hartin, Corinne},
   journal={Plos one},
   volume={14},
@@ -33,7 +33,7 @@
   doi = {10.1371/journal.pone.0223542}
 }
 @article{calvin2019gcam,
-  title={GCAM v5. 1: representing the linkages between energy, water, land, climate, and economic systems},
+  title={GCAM v5.1: representing the linkages between energy, water, land, climate, and economic systems},
   author={Calvin, Katherine and Patel, Pralit and Clarke, Leon and Asrar, Ghassem and Bond-Lamberty, Ben and Cui, Ryna Yiyun and Di Vittorio, Alan and Dorheim, Kalyn and Edmonds, Jae and Hartin, Corinne and others},
   journal={Geoscientific Model Development (Online)},
   volume={12},
@@ -77,7 +77,7 @@
   year={2007}
 }
 @article{thomson2011rcp4,
-  title={RCP4. 5: a pathway for stabilization of radiative forcing by 2100},
+  title={RCP4.5: a pathway for stabilization of radiative forcing by 2100},
   author={Thomson, Allison M and Calvin, Katherine V and Smith, Steven J and Kyle, G Page and Volke, April and Patel, Pralit and Delgado-Arias, Sabrina and Bond-Lamberty, Ben and Wise, Marshall A and Clarke, Leon E and others},
   journal={Climatic change},
   volume={109},
@@ -88,7 +88,7 @@
   doi = {10.1007/s10584-011-0151-4}
 }
 @article{van2011rcp2,
-  title={RCP2. 6: exploring the possibility to keep global mean temperature increase below 2 C},
+  title={RCP2.6: exploring the possibility to keep global mean temperature increase below 2${}^\circ$C},
   author={Van Vuuren, Detlef P and Stehfest, Elke and den Elzen, Michel GJ and Kram, Tom and van Vliet, Jasper and Deetman, Sebastiaan and Isaac, Morna and Goldewijk, Kees Klein and Hof, Andries and Beltran, Angelica Mendoza and others},
   journal={Climatic Change},
   volume={109},
@@ -99,7 +99,7 @@
   doi = {10.1007/s10584-011-0152-3}
 }
 @Manual{shiny,
-  title = {shiny: Web Application Framework for R},
+  title = {shiny: Web Application Framework for {R}},
   author = {Winston Chang and Joe Cheng and JJ Allaire and Yihui Xie and Jonathan McPherson},
   year = {2020},
   note = {R package version 1.4.0.2},
