### 2020-12-08 72fc7bd
make celsius abbrev capitalized
