+++ paper/paper.bib
@@ -88,7 +88,7 @@
   doi = {10.1007/s10584-011-0151-4}
 }
 @article{van2011rcp2,
-  title={RCP2.6: exploring the possibility to keep global mean temperature increase below 2${}^\circ$C},
+  title={RCP2.6: exploring the possibility to keep global mean temperature increase below 2${}^\circ${C}},
   author={Van Vuuren, Detlef P and Stehfest, Elke and den Elzen, Michel GJ and Kram, Tom and van Vliet, Jasper and Deetman, Sebastiaan and Isaac, Morna and Goldewijk, Kees Klein and Hof, Andries and Beltran, Angelica Mendoza and others},
   journal={Climatic Change},
   volume={109},
