### 2020-12-08 #2782 @crvernon
@kthyng I updated the references to be formatted correctly.  Thank you for catching that!  The person who originally linked the `hectorui` repository with Zenodo is no longer with our organization.  This keeps me from being able to edit the metadata of the Zenodo record to update the citation as you requested.  I have contacted Zenodo to see if they can terminate the existing connection so that I may activate the `hectorui` repo on my Zenodo account.  I will let you know when I hear back and have this resolved.  Thanks!
