### 2020-12-08 #2782 @kthyng
@crvernon Ok paper looks good including references, and I'll wait to hear back on the Zenodo metadata. Thanks!
