### 2020-12-11 #2782 @crvernon
Alright @kthyng we have the Zenodo issue worked out. The newest release will is v1.1.6 which has been minted at the following with an updated reference:

Jason Evanoff, Chris R. Vernon, Stephanie Waldhoff, Abigail Snyder, & Corinne Hartin. (2020, December 11). JGCRI/hectorui: v1.1.6 JOSS publication (Version v1.1.6). Zenodo. http://doi.org/10.5281/zenodo.4317756

Thanks for everything!
