### 2020-12-11 #2782 @kthyng
@crvernon Good, though can you make the title "hectorui: A web-based interactive scenario builder and visualization application for the Hector climate model"?
