### 2020-12-11 #2782 @crvernon
No problem. Just updated it. Thanks!

Jason Evanoff, Chris R. Vernon, Stephanie Waldhoff, Abigail Snyder, & Corinne Hartin. (2020, December 11). hectorui: A web-based interactive scenario builder and visualization application for the Hector climate model (Version v1.1.6). Zenodo. http://doi.org/10.5281/zenodo.4317756
