### 2020-12-11 #2782 @whedon
:wave: @openjournals/joss-eics, this paper is ready to be accepted and published.
